pyproject.toml
[project]
name = "heygen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-limiter>=0.1.6",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "requests>=2.32.3",
    "sqlmodel>=0.0.22",
    "fakeredis>=2.26.1",
    "uvicorn>=0.32.1",
]
